+++ packages/components/src/components/Menu/index.ts
@@ -8,4 +8,5 @@ type MenuFamily = typeof _Menu & {
 const Menu = _Menu as MenuFamily;
 Menu.Item = MenuItem;
 
-export { Menu, MenuProps, MenuItemProps };
+export { Menu };
+export type { MenuProps, MenuItemProps };